New exercises added

+++ week-1/day-3/Exercises/src/MileToKmConverter.java
@@ -9,7 +9,7 @@ public class MileToKmConverter {
             double L = scan.nextInt();
             scan.close();
 
-            System.out.println("Miles in Km: " + (L/5*8) );
+            System.out.println("Miles in Km: " + ((double)L/5*8));
             
         // Write a program that asks for a double that is a distance in miles,
         // then it converts that value to kilometers and prints it
